moved proxy functionality to core folder
